Add a new exercise to use CSV instead of JSON
Decided to have use CSV second instead of first because it may be easier
to mentally map between JSON and ruby objects for beginners due to the
similar syntax.

Once the concept of mapping between JSON and ruby objects is clear,
introducing CSV to ruby objects may make more sence and the audience
probably has experience with Excel and should be a nice way of showing
how this can be practically used from datasets they already have.